Create GithubCommitAnalyzer.py
The program is menu driven. It can analyse the commits made by students of the Software Engineering class for a particular project. Data mining is done on a text file which is created from an XML file.
The key features of the script were that it could find the Most Active Member of the Team, the Most Regular Member in the Team, the Number of Commits made per Member of the Team, the Number of Days of Inactivity of the Team, and the Log Details by Member Name and analyse the aforementioned data in graphical dashboards.